rename const
