+++ src/galax/potential/_src/builtin/builtin.py
@@ -36,11 +36,11 @@ from unxt.unitsystems import dimensionless, galactic
 from xmmutablemap import ImmutableMap
 
 import galax.typing as gt
-from ._const import _burkert_const
-from galax.potential._src.base import default_constants
-from galax.potential._src.core import AbstractPotential
-from galax.potential._src.params.core import AbstractParameter
-from galax.potential._src.params.field import ParameterField
+from .const import _burkert_const
+from galax.potential._potential.base import default_constants
+from galax.potential._potential.core import AbstractPotential
+from galax.potential._potential.params.core import AbstractParameter
+from galax.potential._potential.params.field import ParameterField
 
 # -------------------------------------------------------------------
