+++ src/galax/potential/_src/builtin/const.py
@@ -1,0 +1,5 @@
+import quaxed.array_api as xp
+
+_burkert_const = 3 * xp.log(xp.asarray(2.0)) - 0.5 * xp.pi
+_log2 = xp.log(xp.asarray(2.0))
+_sqrt2 = xp.sqrt(xp.asarray(2.0))
